Implement peg ordering and sane peg output
Prior to this commit, the nature of the recursive algorithm that solves
the puzzle meant that pegs would be displayed in a different order every
time. In order to allow the order of the pegs to be determined, I had to
implement a number of traits (Ord, PartialOrd, PartialEq) in order to
implement the sorting logic of this arbitrary datatype. It works!