Major update that refactors display logic and types
+ Output now looks somewhat nice :)
+ Note that I did consider wrapping a `Disc` in the Option<T> enum, but
decided to implement my own, to properly convey the correct meaning
(i.e. each Peg is loaded with discs, and/or empty space (not "discs
and/or empty discs" - there really is no such thing as an empty disc).
Yes, the difference is only small, but this approach also allowed me to
make greater use of Rust's enums (read: tagged unions) as a learning
exercise.