Adds argument handling for game size and refresh interval
Note that this commit also introduces a new feature - a global
configuration structure, intended to simplify function signatures,
and to allow me to avoid adding dozens of lifetime specifiers
throughout the codebase

Presently, only the `refresh_interval` struct member is in use,
though a `game_size` is present also. I may remove the latter in
a future commit.